[MOD] Cleanups

+++ src/test/java/org/basex/test/xmldb/CollectionTest.java
@@ -3,8 +3,10 @@ package org.basex.test.xmldb;
 import static org.junit.Assert.*;
 
 import java.io.File;
+
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -21,7 +23,6 @@ import org.xmldb.api.base.XMLDBException;
 import org.xmldb.api.modules.BinaryResource;
 import org.xmldb.api.modules.XMLResource;
 import org.xmldb.api.modules.XPathQueryService;
-import junit.framework.TestCase;
 
 /**
  * This class tests the XMLDB/API Collection implementation.
@@ -378,7 +379,7 @@ public class CollectionTest extends XMLDBBaseTest {
      * @throws XMLDBException exception
      */
     int num() throws XMLDBException {
-      return ((Integer) run()).intValue();
+      return (Integer) run();
     }
 
     /**
